SCRUM-42 Added land use data to DataFetcher
Added grabbing of land usage.

It is in format of area (geometry) with type property (type enum needs to be implemented).

To get the type of land lot a geometry intersection query with land type data should be performed. The result with most area overlap is the land lot type.