SCRUM-60, SCRUM-61, SCRUM-62 UI improvements (details in description)
Sidebar - The styles were in the same file as the component itself so i moved them to the Styles.js file for consistency. I also improved the overall look and feel of the sidebar adding icons from the Figma mockups and  smother animations. The sidebar is now also static and always takes up the same amount of space and does not expand with the content displayed on the right.

Smart Select and Side By Side - I added the map and result bar components to these pages.

Search Bar - I implemented a search bar component that will allow us to search for lots and/or addresses one the backend functionality is in place. For now it is added to Assess page and just sends its content into console upon user pressing enter.