SCRUM-73 Finished Parser
Implemented functions for all productions specified in BNF.
Each production returns an AST node or throws an exception in case of an invalid token.
The final result is an AST tree or null in case of an error.

Fixed BNF as some productions did not align with the language design in mind.
There are still some limitations in current BNF that need to be addressed.
The most annoying one being that lambda return values cannot contain numeric expressions, so a constant has to be declared and then returned (but this requires a runtime hack)