SCRUM-100 Added Heat-maps to Risk Lens
Created flood heatmap data as a collection of GeoJSON points where each point has a score.
Data was obtained by splitting the whole Slovenia BBOX into 392x256 grid and performing an
intersection test with flood areas in DB for each grid cell. Score was given based on the severity of the intersected flood areas.

Flood data is saved in /public folder in backend as generation takes quite a bit of time and manual effort (~30min).

Created landslide heatmap dataset from image raster obtained from API from which landslide geometry data was obtained.
Tried using the same program / method as was used for floods but some database intersection queries would take 30+ seconds which proved unfeasable for
a grid with over 100k points.
As a fallback I wrote a program for processing raster images generated by ArcGis API where each pixel is color coded.

Dataset was generated from an image with dimensions of 6252px x 4100px
where a block of 6x4 pixels was sampled to determine the final score in the dataset point (sum of individual pixel values).

NOTE: current dataset is a bit large (~22mb) and can be regenerated if necessary (left some config options in frontend for it aswell)
although this reduces the level of detail quite a bit as a larger block is sampled and averaged for each data point.

This data is also saved in /public folder.

Earthquake data is retrieved from the database dynamically as it is the only one prone to change, is already stored in form of points
and has a low record count.

As an addition when viewing earthquakes, points of epicenters are added to the map and information about them is displayed
when the mouse cursor is near (magnitude, timestamp).
Nearby earthquakes are grouped together in the popup (20px grouping radius) to improve clarity.

Added MapBox layers for each heatmap in Risk Lens tab. Functionality for each layer was split into a seperate files
as it contains quite a bit of code and would make the main Map.js module cluttered.

Heatmap data stays consistent up untill zoom level 14 and fades out at zoom level 15 to not block undelaying land lots.
(At least I tried to make it as consistent with zoom by providing heatmap radius values for each zoom level the heatmap is visible).

Color values for heatmaps can be adjusted if desired.