SCRUM-108 Added context menu and clipping mask
Context menu is shown when right-clicking on a land lot. The lot is marked as selected (outline)
and a clipping mask shadows all other content except the lot itself.

It is hidden if the close button or any other area besides the menu itself is clicked.

Clipping mask color and speed can be customized.
Context menu can be expanded as it currenlty only has two options (save lot, add to comparison)