SCRUM-108 Managing and viewing saved lots on map
Lots can now be saved. A name and address must be supplied (address is reverse geocoded and automatically supplied. It can be edited if invalid / not found).
Saved lots are stored in DB per user. One save per land lot limit is currently inforced (might change this in the future).
Saves can be removed or vieved on map. When viewing the saved lot on map, it is zoomed to and highlighted on load which enables URL sharing.

Also fixed some UserContext things as it wasn't initialized properly causing requests to fail due to invalid auth (should probably be rewritten).