SCRUM-30 Partial Smart-Select implementation
Points can be added with a right click and removed with left click.
Left-clicking near a point (10px radius) removes just that one point, while if no point is near, the whole selection is reset.
Right-clicking near the first point completes the selection.

Once the area is selected its gets sent to backend for computation (filters are yet to be implemented).

TODO:
    -Panning the map adds unwanted points to selection.
    -Implement nicer UI and add selection filters
    -Improve assessment performance (current ~50ms/land lot; which takes A WHILE for a bit larger area)